backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.27.0
email-validator>=2.0.0
itsdangerous>=2.1.0
python-multipart>=0.0.9
motor>=3.6.0
openai>=1.0.0
groq>=1.0.0